Move manual entries to history
