+++ app/res/drawable/fab_compat_bg.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <solid android:color="@color/colorPrimaryFab" />
+    <corners android:radius="28dp" />
+</shape>
