+++ app/res/layout-v15/history.xml
@@ -1,0 +1,47 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<!--
+      ~ Copyright (C) 2012 jonas.oreland@gmail.com
+      ~
+      ~  This program is free software: you can redistribute it and/or modify
+      ~  it under the terms of the GNU General Public License as published by
+      ~  the Free Software Foundation, either version 3 of the License, or
+      ~  (at your option) any later version.
+      ~
+      ~  This program is distributed in the hope that it will be useful,
+      ~  but WITHOUT ANY WARRANTY; without even the implied warranty of
+      ~  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+      ~  GNU General Public License for more details.
+      ~
+      ~  You should have received a copy of the GNU General Public License
+      ~  along with this program.  If not, see <http://www.gnu.org/licenses/>.
+      -->
+
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/tab1_layout"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent">
+
+    <ListView
+        android:id="@+id/history_list"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:clipToPadding="false"
+        android:paddingBottom="@dimen/margin_for_fab"
+        android:scrollbarStyle="outsideOverlay" />
+
+    <android.support.design.widget.FloatingActionButton
+        android:id="@+id/history_add"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="right|bottom"
+        android:layout_margin="16dp"
+        android:contentDescription="@string/Add_manual_entry"
+        android:tooltipText="@string/Add_manual_entry"
+        app:backgroundTint="@color/colorPrimaryFab"
+        app:srcCompat="@drawable/ic_add_white_24dp"
+        tools:ignore="UnusedAttribute" />
+
+</FrameLayout>
