+++ app/res/layout/title_spinner.xml
@@ -18,7 +18,8 @@
     android:id="@+id/title_spinner"
     android:layout_width="fill_parent"
     android:layout_height="fill_parent"
-    android:layout_margin="4dp"
+    android:layout_marginBottom="4dp"
+    android:layout_marginTop="4dp"
     android:background="@drawable/title_spinner"
     android:clickable="true"
     android:gravity="center_vertical"
