+++ app/res/menu/manual_menu.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:id="@+id/menu_save"
+        android:orderInCategory="100"
+        app:showAsAction="always"
+        android:title="@string/Save" />
+</menu>
