+++ app/res/values/colors.xml
@@ -24,6 +24,8 @@
      <color name="colorText">#ffffff</color>
      <color name="colorTextSecondary">#b3ffffff</color>
 
+     <color name="colorPrimaryFab">#0066B7</color>
+
      <color name="serviceDefault">#003366</color>
      <color name="serviceDigifit">#5DA9D3</color>
      <color name="serviceEndomondo">#87AA14</color>
