+++ app/res/values/dimens.xml
@@ -6,5 +6,6 @@ Refer to App Widget Documentation for margin information
 http://developer.android.com/guide/topics/appwidgets/index.html#CreatingLayout
     -->
     <dimen name="widget_margin">8dp</dimen>
-
+    <dimen name="margin_for_fab">88dp</dimen>
+    <dimen name="activity_margin">16dp</dimen>
 </resources>
