+++ common/src/main/res/values/strings.xml
@@ -94,7 +94,7 @@
   <string name="High_pace">High pace</string>
   <string name="Sex">Sex</string>
   <string name="upload">upload</string>
-  <string name="feed">feed</string>
+  <string name="feed">Feed</string>
   <string name="Notes">Notes</string>
   <string name="Laps">Laps</string>
   <string name="Map">Map</string>
@@ -249,6 +249,8 @@
   <string name="Basic">Basic</string>
   <string name="Advanced">Advanced</string>
   <string name="Manual">Manual</string>
+  <string name="Manual_activity_title">Add manual entry</string>
+  <string name="Add_manual_entry">Add manual entry</string>
   <string name="Add_workout_notes">Add workout notes</string>
   <string name="Pruning_deleted_activities_from_database">Pruning deleted activities from database</string>
   <string name="Prune">Prune</string>
@@ -289,6 +291,7 @@
   <string name="HinHere_is_a_workout_I_think_you_might_like">Hi\nHere is a workout I think you might like.</string>
   <string name="Share_workout">Share workout…</string>
   <string name="New_audio_scheme">New audio scheme</string>
+  <string name="GPS_permission_required">GPS permission is required</string>
   <string name="Default">Default</string>
   <string name="Notes_about_your_workout">Notes about your workout</string>
   <string name="Downloading_from_1s">Downloading from %1$s</string>
